requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pillow==10.3.0